Feat/mamuri

See merge request s03-final/s03p31a210!341

+++ frontend/src/components/TipDetail.vue
@@ -1,43 +1,33 @@
 <template>
   <div class="tip-detail">
-    <table class="table table-striped table-bordered table-hover mt-3">
+    <h3 style="text-align: left; padding-left: 10%;">{{notice.category}}</h3>
+    <table class="table table-bordered mt-3" style="width: 80%; margin: auto; font-size: 15px;">
       <thead>
-        <tr>
-          <th>제목</th>
-          <th>{{notice.title}}</th>
+        <tr style="background-color: #bbbfca;">
+          <th rowspan="2" style="width: 70%; vertical-align: middle;">{{notice.title}}</th>
+          <th style="width: 20%;">
+            <span v-if="notice === {}">{{notice.date.slice(0,10)}}</span>
+          </th>
+          <th style="width: 10%;">조회 {{notice.hits}}</th>
+        </tr>
+        <tr style="background-color: #bbbfca;">
+          <th colspan="2">{{notice.name}}</th>
         </tr>
       </thead>
       <tbody>
         <tr>
-          <td>카테고리</td>
-          <td style="white-space:pre;">{{notice.category}}</td>
-        </tr>
-        <tr>
-          <td>작성자</td>
-          <td style="white-space:pre;">{{notice.name}}</td>
-        </tr>
-        <tr>
-          <td>날짜</td>
-          <td style="white-space:pre;">{{notice.date.slice(0,10)}}</td>
-        </tr>
-        <tr>
-          <td>내용</td>
-          <td style="white-space:pre;">{{notice.content}}</td>
-        </tr>
-        <tr>
-          <td>조회수</td>
-          <td style="white-space:pre;">{{notice.hits}}</td>
+          <td colspan="4" style="padding: 0;">
+            <b-textarea disabled no-resize :value="notice.content" style="height: 400px; background-color: #e8e8e8;"/>
+          </td>
         </tr>
       </tbody>
     </table>
     
-    <!-- <b-button v-if="notice.user == user" class="mybutton1" @click="moveNoticeModify()">수정</b-button> -->
-    <b-button v-if="notice.user == yol_uid" class="mybutton1" @click="moveNoticeModify()">수정</b-button>
+    <b-button v-if="notice.user == yol_uid" class="mybutton1" @click="moveNoticeModify()" style="margin: 20px;">수정</b-button>
 
-    <!-- <b-button  v-if="notice.user == user" class="mybutton2 mx-3" @click="moveNoticeDelete()">삭제</b-button> -->
-    <b-button v-if="notice.user == yol_uid" class="mybutton2 mx-3" @click="moveNoticeDelete()">삭제</b-button>
+    <b-button v-if="notice.user == yol_uid" class="mybutton2 mx-3" @click="moveNoticeDelete()" style="margin: 20px;">삭제</b-button>
 
-    <b-button class="border-0" @click="moveNotice()">목록으로</b-button>
+    <b-button class="border-0" @click="moveNotice()" style="margin: 20px;">목록으로</b-button>
 
     <div class="d-flex justify-content-between">
       <h4 class="my-3 text-left">댓글 <b-icon-chat-text></b-icon-chat-text></h4>
@@ -58,31 +48,21 @@
         </b-list-group-item>
       </b-list-group>
 
-    <b-modal ref="mdAddCmt" hide-footer title="댓글 작성" >
+    <b-modal ref="mdAddCmt" hide-footer hide-header>
       <b-form @submit="addCmt">
-        <!-- <b-form-group label="작성자" label-for="f-a-c-content">
-          <b-form-textarea
-            id="f-a-c-content"
-            v-model="replier"
-            placeholder="작성자를 입력해주세요"
-          ></b-form-textarea>
-        </b-form-group> -->
-        <!-- </b-form-group> 
-            id="f-m-c-ontent" 
-            v-model="replier" 
-            placeholder="작성자를 입력해주세요"
-          ></b-form-text> -->
         <b-form-group label="댓글" label-for="f-a-c-content">
           <b-form-textarea
+            no-resize
             id="f-a-c-content"
             v-model="replycontent"
             placeholder="댓글을 입력해주세요"
             :rows="1"
             :max-rows="2"
+            style="height: 100px;"
           ></b-form-textarea>
         </b-form-group>
 
-        <b-btn type="submit" variant="primary" class="float-right">댓글 쓰기</b-btn>
+        <b-button type="submit" class="float-right" style="background-color: #084481;">댓글 쓰기</b-button>
       </b-form>
     </b-modal>
     
@@ -152,9 +132,7 @@ export default {
     // ...mapGetters(['userinfo', 'isLogin']),
   },
   created() {
-    const uid = this.$cookies.get("yol_uid");
-    console.log(uid);
-
+    window.scrollTo(0,0);
 
     axios({
       method: "GET",
@@ -165,7 +143,6 @@ export default {
     //   },
     })
       .then(({ data }) => {
-        console.log(data);
         this.notice = data.data;
         
       })
@@ -180,7 +157,6 @@ export default {
 
     })
       .then(({ data }) =>{
-        console.log(data);
         this.comments=data.data;
       })
 
@@ -221,6 +197,11 @@ export default {
 
       const uid = this.$cookies.get("yol_uid");
 
+     if(this.replycontent === null || this.replycontent.length === 0){
+       alert("내용을 입력해주세요!");
+       return;
+     }
+
       axios({
         method: "POST",
         url:
@@ -235,12 +216,7 @@ export default {
         // },
       })
         .then(({ data }) => {
-          console.log(data);
-          console.log(data.comments);
           this.comments = data.comments;
-          console.log(this.comments);
-          // this.datas=data;
-          // console.log(this.datas)
           return this.swalSuccess("댓글 추가 완료");
 
           // if (data == 'success'){
@@ -285,9 +261,6 @@ export default {
         .then((data) => {
           if (!data) throw new Error("");
 
-          // console.log(token);
-          // console.log(333);
-          // console.log(this.commenter.id);
           axios({
             method: "PUT",
             url:
@@ -300,8 +273,7 @@ export default {
             //   Authorization: `Token ${token}`,
             // },
           })
-            .then((res) => {
-              console.log(res.data);
+            .then(() => {
             })
             .catch((err) => {
               console.log(err);
@@ -331,7 +303,6 @@ export default {
       // this.formCmt.id = v.id;
       // this.formCmt.content = v.content;
       // const token = this.$cookies.get("auth-token");
-      // console.log(token);
 
       axios({
         method: "GET",
@@ -342,10 +313,7 @@ export default {
         // },
       })
         .then((res) => {
-          console.log(res.data);
           this.commenter = res.data.data;
-          // console.log(res.data.id);
-          // this.find_pk = res.data.id;
         })
         .catch((err) => {
           console.log(err);
@@ -363,23 +331,19 @@ export default {
       });
     },
     moveNoticeDelete() {
-    //   const token = this.$cookies.get("auth-token");
-    //   console.log(token);
-    //   console.log();
-      axios({
-        method: "DELETE",
-        url: `${API_URL}/tip/delete/` + this.$route.params.id,
-        // headers: {
-        //   Authorization: `Token ${token}`,
-        // },
-      })
-        .then(() => {
-          this.moveNotice();
+      if(confirm("정말로 이 글을 삭제하시겠습니까?")){
+        axios({
+          method: "DELETE",
+          url: `${API_URL}/tip/delete/` + this.$route.params.id,
         })
-        .catch((err) => {
-          console.log(err);
-          alert("삭제 에러가 발생했습니다.");
-        });
+          .then(() => {
+            this.moveNotice();
+          })
+          .catch((err) => {
+            console.log(err);
+            alert("삭제 에러가 발생했습니다.");
+          });
+      }
     },
     delCmt(cmt_id) {
       this.$swal({
@@ -396,11 +360,6 @@ export default {
         },
       })
         .then(() => {
-          // if (!res) throw new Error('');
-          // const token = this.$cookies.get("auth-token");
-          // console.log(token);
-          // console.log(333);
-          // console.log(this.commenter.id);
           axios({
             method: "DELETE",
             url:
@@ -410,8 +369,7 @@ export default {
             //   Authorization: `Token ${token}`,
             // },
           })
-            .then((res) => {
-              console.log(res.data);
+            .then(() => {
             })
             .catch((err) => {
               console.log(err);
@@ -452,7 +410,7 @@ export default {
   }
   .tip-detail {
     padding-top: 100px;
-    background-color: #F2F2F5;
+    background-color: #f4f4f2;
     padding-left: 15%;
     padding-right: 15%;
     padding-bottom: 5%;
